+++ playwright.ui.config.ts
@@ -6,7 +6,7 @@ import { defineConfig } from '@playwright/test'
  */
 export default defineConfig({
   testDir: './tests/e2e',
-  testMatch: 'ui-layout.spec.ts',
+  testMatch: ['ui-layout.spec.ts', 'playground-layout.spec.ts'],
   workers: 1,
   forbidOnly: Boolean(process.env.CI),
   retries: 0,